Last Dump Of All The Recent Changes/Additions
Summary:
Last commit with all the changes from Dec through Jan.
Adds profile with tapglue api data - tests for retrieving data
Adds recipe idea list screen - (hooked up to parse)
Updates food tracking / calorie counter
Changes to workout countdown (may not be working correctly anymore)

That's it!

+++ app/src/main/java/com/femlite/app/model/Portion.java
@@ -9,5 +9,5 @@ public interface Portion {
 
     int getCalories();
 
-    int getServingSize();
+    double getServingSize();
 }
